V1.0.1-Alpha release
Changelog

- Ablation is now functional. One can click the nodes on the graph while
holding shift key to deactivate/reactivate the neurons. Deactivaing
certain neuron zerorizes both row and column for that neuron in gap and
chemical connectome
- The ablation can be done during the simulation as well but this may
result in error in python integrator. For now, I recommend modifying the
connectome BEFORE you start the simulation.  Ablation interface for the
neuron label will be added in the next release.
- The label while hovering the nodes is now sorted in alphabetical order
- Debug mode has been added for the developers